python-service/requirements.txt
dapr
dapr-ext-fastapi
dapr-ext-grpc
fastapi
google-generativeai
httpx
markdown
pusher
python-dotenv
uvicorn[standard]